+++ packages/keyring-controller/src/index.ts
@@ -1,3 +1,4 @@
 export * from './KeyringController';
 export type * from './types';
 export * from './errors';
+export { KeyringControllerErrorMessage } from './constants';